[pvr] remove libXBMC_codec.h and add to libXBMC_pvr.h
This remove the use of libXBMC_codec.h where was nearly nothing
in and was only needed for PVR add-ons. Now becomes the parts
of them moved to the PVR headers.

Next step of the addon system rework.
