--- addons/xbmc.codec/addon.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<addon id="xbmc.codec" version="1.0.1" provider-name="Team Kodi">
-  <backwards-compatibility abi="1.0.1"/>
-  <requires>
-    <import addon="xbmc.core" version="0.1.0"/>
-  </requires>
-</addon>
