+++ addons/xbmc.pvr/addon.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<addon id="xbmc.pvr" version="5.2.1" provider-name="Team-Kodi">
-  <backwards-compatibility abi="5.2.1"/>
+<addon id="xbmc.pvr" version="5.2.2" provider-name="Team-Kodi">
+  <backwards-compatibility abi="5.2.2"/>
   <requires>
     <import addon="xbmc.core" version="0.1.0"/>
   </requires>
