--- xbmc/addons/interfaces/Codec/CMakeLists.txt
@@ -1,5 +1,0 @@
-set(SOURCES AddonCallbacksCodec.cpp)
-
-set(HEADERS AddonCallbacksCodec.h)
-
-core_add_library(addonInterface_Codec)
