+++ xbmc/addons/interfaces/PVR/AddonCallbacksPVR.cpp
@@ -39,6 +39,10 @@
 #include "utils/log.h"
 #include "utils/StringUtils.h"
 
+extern "C" {
+#include "libavcodec/avcodec.h"
+}
+
 using namespace ADDON;
 using namespace PVR;
 
@@ -46,6 +50,67 @@ namespace KodiAPI
 {
 namespace PVR
 {
+
+class CCodecIds
+{
+public:
+  virtual ~CCodecIds(void) {}
+
+  static CCodecIds& GetInstance()
+  {
+    static CCodecIds _instance;
+    return _instance;
+  }
+
+  xbmc_codec_t GetCodecByName(const char* strCodecName)
+  {
+    xbmc_codec_t retVal = XBMC_INVALID_CODEC;
+    if (strlen(strCodecName) == 0)
+      return retVal;
+
+    std::string strUpperCodecName = strCodecName;
+    StringUtils::ToUpper(strUpperCodecName);
+
+    std::map<std::string, xbmc_codec_t>::const_iterator it = m_lookup.find(strUpperCodecName);
+    if (it != m_lookup.end())
+      retVal = it->second;
+
+    return retVal;
+  }
+
+private:
+  CCodecIds(void)
+  {
+    // get ids and names
+    AVCodec* codec = NULL;
+    xbmc_codec_t tmp;
+    while ((codec = av_codec_next(codec)))
+    {
+      if (av_codec_is_decoder(codec))
+      {
+        tmp.codec_type = (xbmc_codec_type_t)codec->type;
+        tmp.codec_id   = codec->id;
+
+        std::string strUpperCodecName = codec->name;
+        StringUtils::ToUpper(strUpperCodecName);
+
+        m_lookup.insert(std::make_pair(strUpperCodecName, tmp));
+      }
+    }
+
+    // teletext is not returned by av_codec_next. we got our own decoder
+    tmp.codec_type = XBMC_CODEC_TYPE_SUBTITLE;
+    tmp.codec_id   = AV_CODEC_ID_DVB_TELETEXT;
+    m_lookup.insert(std::make_pair("TELETEXT", tmp));
+
+    // rds is not returned by av_codec_next. we got our own decoder
+    tmp.codec_type = XBMC_CODEC_TYPE_RDS;
+    tmp.codec_id   = AV_CODEC_ID_NONE;
+    m_lookup.insert(std::make_pair("RDS", tmp));
+  }
+
+  std::map<std::string, xbmc_codec_t> m_lookup;
+};
 
 CAddonCallbacksPVR::CAddonCallbacksPVR(CAddon* addon)
   : m_addon(addon),
@@ -69,6 +134,7 @@ CAddonCallbacksPVR::CAddonCallbacksPVR(CAddon* addon)
   m_callbacks->TransferChannelGroupMember = PVRTransferChannelGroupMember;
   m_callbacks->ConnectionStateChange      = PVRConnectionStateChange;
   m_callbacks->EpgEventStateChange        = PVREpgEventStateChange;
+  m_callbacks->GetCodecByName = GetCodecByName;
 }
 
 CAddonCallbacksPVR::~CAddonCallbacksPVR()
@@ -421,6 +487,12 @@ void CAddonCallbacksPVR::PVREpgEventStateChange(void* addonData, EPG_TAG* tag, u
     CSingleLock lock(queueMutex);
     queuedChanges.push_back(EpgEventStateChange(client->GetID(), iUniqueChannelId, tag, newState));
   }
+}
+
+xbmc_codec_t CAddonCallbacksPVR::GetCodecByName(const void* addonData, const char* strCodecName)
+{
+  (void)addonData;
+  return CCodecIds::GetInstance().GetCodecByName(strCodecName);
 }
 
 } /* namespace PVR */
